Floating number fix.

+++ command/gps.go
@@ -29,9 +29,9 @@ func processLocation(content []byte) int {
 
 	log.Printf("GPS_INFORMATION %d", content[6])
 	log.Printf("Number of satelites %d", content[7])
-	latitude := util.BytesToInt(content[7:11]) / 1800000
-	longitude := util.BytesToInt(content[11:15]) / 18000000
-	log.Printf("Location %d,%d", latitude, longitude)
+	latitude := float64(util.BytesToInt(content[7:11])) / 1800000
+	longitude := float64(util.BytesToInt(content[11:15])) / 18000000
+	log.Printf("Location %.7f,%.7f", latitude, longitude)
 
 	return 16
